Insider Threat Detection/requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
joblib
psycopg2
python-dotenv
